patch-id: fix scan_hunk_header on diffs with 1 line of before/after
The "diff" parser in "patch-id" forgot that the hunks, which usually
looks like "@@ -2,2 +2,15 @@ code", may not have the ",<number>" if
the number of lines in the preimage or the postimage is exactly 1,
e.g. "@@ -2 +2,18 @@ code" or "@@ -2,2 +2 @@ code".

As a result, scan_hunk_header mistakenly returns the line
number as line count, which then results in unpredictable
parsing errors with the rest of the patch, including giving
multiple lines of output for a single commit.

Fix by explicitly setting line count to 1 when there is
no comma, and add a test.

apply.c contains this same logic except it is correct. A
worthwhile future project might be to unify these two diff
parsers so they both benefit from fixes.

Signed-off-by: Jerry Zhang <jerry@skydio.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>