Merge with untracked file that are the same without failure and test
When doing a merge while there is untracked files with the same name
as merged files, git refuses to proceed. This commit change this
behavior and make git overwrite files if their contents are the same.
This new behaviour is more pleasant for a user and will never be a 
frustrating moment.

Add a if statement that check if the file has the same content as the 
merged file thanks to the function ie_modified() (read-cache.c).
ie_modified () checks the status of both files, if they are different, 
it verifies their contents.

Add new tests that need to pass to confirm that the new feature works.

Co-authored-by: COGONI Guillaume <cogoni.guillaume@gmail.com>