merge with untracked file that are the same without failure
Keep the old behaviour as default.

Add the option --overwrite-same-content.
When this option is used merge will overwrite untracked file that have
the same content.

In unpack-trees.c in the check_ok_to_remove() function: add a new
statement, if the file has the same content as the merged file, it
can be overwrited.

t7615 tests this new behaviour.

Signed-off-by: Jonathan Bressat <git.jonathan.bressat@gmail.com>
Signed-off-by: COGONI Guillaume <cogoni.guillaume@gmail.com>